Get rid of Sinatra namespace
