+++ lib/stoplight_admin/actions/home.rb
@@ -1,28 +1,26 @@
 # frozen_string_literal: true
 
-require 'sinatra/stoplight_admin/lights_repository'
-
-module Sinatra
-  module StoplightAdmin
-    class HomeAction
+module StoplightAdmin
+  module Actions
+    class Home
       # @!attribute lights_repository
-      #   @return [Sinatra::StoplightAdmin::LightsRepository]
+      #   @return [StoplightAdmin::LightsRepository]
       attr_reader :lights_repository
       private :lights_repository
 
       # @!attribute lights_stats
-      #   @return [Sinatra::StoplightAdmin::LightsStats]
+      #   @return [StoplightAdmin::LightsStats]
       attr_reader :lights_stats
       private :lights_stats
 
-      # @return lights_repository [Sinatra::StoplightAdmin::LightsRepository]
-      # @param lights_stats [Class<Sinatra::StoplightAdmin::LightsStats>]
+      # @return lights_repository [StoplightAdmin::LightsRepository]
+      # @param lights_stats [Class<StoplightAdmin::LightsStats>]
       def initialize(lights_repository:, lights_stats:)
         @lights_repository = lights_repository
         @lights_stats = lights_stats
       end
 
-      # @return [(Sinatra::StoplightAdmin::LightsRepository::Light)]
+      # @return [(StoplightAdmin::LightsRepository::Light)]
       def call
         lights = lights_repository.all
         stats = lights_stats.call(lights)
