+++ lib/stoplight_admin/actions/stats.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+module StoplightAdmin
+  module Actions
+    class Stats < Home
+    end
+  end
+end
