+++ lib/stoplight_admin/views/index.erb
