+++ lib/stoplight_admin/views/layout.erb
