+++ spec/support/database_cleaner.rb
@@ -9,7 +9,7 @@ DatabaseCleaner.strategy = cleaning_strategy
 RSpec.shared_context :redis, :redis do
   let(:redis) { Redis.new(url: ENV.fetch('STOPLIGHT_REDIS_URL', 'redis://127.0.0.1:6379/0')) }
 
-  before(:suite) do
+  before do
     DatabaseCleaner[:redis].db = redis
     DatabaseCleaner.clean_with(:deletion)
   end
